+++ carbonj.service/src/main/java/com/demandware/carbonj/service/db/TimeSeriesStoreImpl.java
@@ -358,11 +358,11 @@ public class TimeSeriesStoreImpl implements TimeSeriesStore
     }
 
     @Override
-    public DataPointExportResults exportPoints(String dbName, int metricId) {
+    public DataPointExportResults exportPoints(String dbName, long metricId) {
         return exportPoints(dbName, null, metricId);
     }
 
-    private DataPointExportResults exportPoints(String dbName, String metricName, Integer metricId) {
+    private DataPointExportResults exportPoints(String dbName, String metricName, Long metricId) {
         if (!RetentionPolicy.dbNameExists(dbName)) {
             throw new RuntimeException(String.format("Unknown dbName [%s]", dbName));
         }
@@ -641,13 +641,13 @@ public class TimeSeriesStoreImpl implements TimeSeriesStore
     }
 
     @Override
-    public Metric getMetric( int metricId )
+    public Metric getMetric( long metricId )
     {
         return nameIndex.getMetric( metricId );
     }
 
     @Override
-    public String getMetricName( int metricId )
+    public String getMetricName( long metricId )
     {
         return nameIndex.getMetricName( metricId );
     }
@@ -668,7 +668,7 @@ public class TimeSeriesStoreImpl implements TimeSeriesStore
     }
 
     @Override
-    public int scanMetrics( int start, int end, Consumer<Metric> m )
+    public long scanMetrics( long start, long end, Consumer<Metric> m )
     {
         return nameIndex.scanNames( start, end, m );
     }
